+++ frontend/ui/src/app/app-routing.module.ts
@@ -3,6 +3,7 @@ import { RouterModule, Routes } from '@angular/router';
 import {HomeComponent} from "./components/pages/home/home.component";
 import {CalendarComponent} from "./components/pages/calendar/calendar.component";
 import {PersonComponent} from "./components/pages/society/person/person.component";
+import {RaceComponent} from "./components/pages/history/race/race.component";
 
 const routes: Routes = [
   {
@@ -21,6 +22,10 @@ const routes: Routes = [
   {
     path: "people",
     component: PersonComponent
+  },
+  {
+    path: "races",
+    component: RaceComponent
   }
 ];