+++ frontend/ui/src/app/components/pages/history/race/race.component.ts
@@ -174,4 +174,3 @@ export class RaceComponent implements OnInit {
     );
   }
 }
-